feat: separate seeders into categories (#599)
* feat: separate seeders into categories

* chore: fix linting issues

---------
