+++ apps/drec-api/package.json
@@ -42,8 +42,10 @@
     "test:e2e:single": "sh ./scripts/test-e2e-single.sh",
     "compodoc": "npx compodoc -p tsconfig.json -n \"Drec api documentation\"",
     "coverage:jest": "jest --coverage",
-    "seed": "ts-node seeders",
-    "seed:dummy-data": "ts-node seeders/dummy-data",
+    "seed:dummy-data": "ts-node seeders/executors/dummy-data.ts",
+    "seed:fingerprints": "ts-node seeders/executors/fingerprints.ts",
+    "seed:permissions": "ts-node seeders/executors/permissions.ts",
+    "seed:verifications": "ts-node seeders/executors/verifications.ts",
     "sentry:sourcemaps": "sentry-cli sourcemaps inject --org d-rec --project drec-origin ./dist && sentry-cli sourcemaps upload --org d-rec --project drec-origin ./dist",
     "release": "standard-version",
     "release:minor": "standard-version --release-as minor",
