+++ apps/drec-api/seeders/executors/fingerprints.ts
@@ -1,0 +1,4 @@
+import { SeederExecutor } from '../core/seeder-executor';
+import { FingerprintSeeder } from '../fixtures/fingerprint.seeder';
+
+SeederExecutor.run([FingerprintSeeder]);
