+++ apps/drec-api/seeders/executors/permissions.ts
@@ -1,0 +1,4 @@
+import { SeederExecutor } from '../core/seeder-executor';
+import { PermissionsSeeder } from '../fixtures/permissions.seeder';
+
+SeederExecutor.run([PermissionsSeeder]);
