+++ apps/drec-api/seeders/fixtures/verify-phone-number.seeder.ts
@@ -1,8 +1,8 @@
 import { Repository } from 'typeorm';
-import { SeederInterface } from './seeder-interface';
+import { SeederInterface } from '../core/seeder-interface';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Injectable } from '@nestjs/common';
-import { User } from '../src/pods/user/user.entity';
+import { User } from '../../src/pods/user/user.entity';
 
 @Injectable()
 export class VerifyPhoneNumbersSeeder implements SeederInterface {
